Implementando cambios de la estructura
