+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import AppBar from "./Components/Appbar";
-import ImageCarousel from "./Components/ImageCarousel";
+import AppBar from "./Components/AppBar/Appbar";
+import ImageCarousel from "./Components/ImageCarousel/ImageCarousel";
 import "./Styles/AppBar.css";
-import ShopPage from "./Components/ShopPage";
-import { CartProvider } from "./Context/CartContext";
-import Cart from "./Components/Cart";
-import PaymentPage from "./Components/Payment"; // Importa el componente de pago
+import ShopPage from "./Pages/ShopPage";
+import { CartProvider } from "./Contexts/CartContext";
+import Cart from "./Pages/Cart";
+import PaymentPage from "./Pages/Payment"; // Importa el componente de pago
 
 const App: React.FC = () => {
   const [searchTerm, setSearchTerm] = useState("");
