+++ src/Pages/Payment.tsx
@@ -1,6 +1,6 @@
 import React, { useState, FormEvent } from "react";
 import { useNavigate } from "react-router-dom";
-import { useCart } from "../Context/CartContext";
+import { useCart } from "../Contexts/CartContext";
 import { db } from "../../firebaseConfig"; // Asegúrate de importar tu configuración de Firebase
 import { doc, updateDoc, increment } from "firebase/firestore";
 import "../Styles/AppBar.css";
